refactor: move findUnattachedPVCs to own file
